Adds two tests for users
This commit adds two tests: a list of users with that can sudo,
and a list of users which have the same UID.  The sudo test will
fail if any user has sudo with NOPASSWD - which enables them to
run commands as root without entering a password.  If no users
have sudo with NOPASSWD, the test prints a list of users that can
sudo.

The second test is a list of users with the same UID.  The test
fails if a user other than 'root' has UID 0.  If there aren't
multiple users with UID 0, the test lists all users which have
duplicate UIDs.

Change-Id: I8f9de505604d80ae439c2a3de4884f959a17e394