Adding tests for group ID and name uniqueness, misc fixes
This commit adds a check that the group names and user IDs are
unique.  If the group ID 0 or the group name root are not unique,
the test fails.  Otherwise, the test lists all duplciated group
names and IDs.

This commit also fixes an issue that was causing a test in
test_mem not to run (try with no except), and the passwordless
user test (wrong attribute name in spwd lib).

Change-Id: I0d68279d04d9c9fc791e910c22fc42b4897b0524