Adding a test for running services
This commit adds a test for services which are running on a
system.  Services can be marked as expected running, in which
case the test will fail or warn if one or all of the services are
not running.  Tests can also be marked as expected not running,
in which case the test will fail or warn if one or all of the
services are running.

This commit also removes @test_class.group_test, which no longer
does anything.

Change-Id: I74ddd4a765e98a2b56790c708213b259f69e85a0