Adding HTML report output
This commit adds the first revision of HTML reporting.  To output
HTML results, Recon uses a template which is specified in a config
file.  Recon will build HTML representing results, and then insert
it into the report template in a predefined section $$$RESULTS$$$.

Later more information like the % of tests which passed, failed,
pretty charts, links to explanations and more can be added.

Change-Id: I41470e648b8621d16ddd3520926973287bf0e7d2