Adding a test for secure service config and fix config search
This commit adds a test which examines service configuration files
and alerts on any values which are in specified allowed or
disallowed values.  This test is configurable by providing a
descriptive name, the configuration file to check, and one or more
settings to examine.  For each setting you can specify allowed or
disallowed values, which will determine the pass result of the
test.

This commit also fixes the test utility to search for options in
config files.  It was using ConfigParser before, which ended up
being a broken solution.  Now it implements a simple parser, and
can be customized to use specified comment delimiters and a
specified key/value delimeter.

Finally, this fixes a small bug in the display of group results.
Previously a parent would be displayed if it passed, when it
shouldn't have.

Change-Id: I56eb2952b7a3010bba814de99cf85a1f8551cb8d