Adding meaningful exit codes
This changes Recon behavior slightly so that exit codes are
meaninful.  Namely, exit code of 2 indicates an error in running
Recon.  Exit code of 1 means that one or more tests failed.  Exit
code of 0 means that all run tests passed.

This is required to support automation in gates and other uses.

Change-Id: Id7ac8789349d461086fe490ee334d22d6c2434a5