Add capability to create a self-contained profile directory
This commit adds the ability to point RBF at a profile directory
which contains config, scripts, and potentially additional
modules.  A profile must contain rbf.cfg as well as the modules,
templates, scripts, and config directories.  For scripts, module
configs, and templates the profile version will take precedence.

If the profile version a referenced file isn't found, Recon will
try to use the version included in the base Recon config.

Change-Id: I5cecfc541a6ee39a345840d3a97c70bf20d3e543