Fixing terminal output to look neater
This commit changes terminal output so that if a test name exceeds
the field width for a test name, the results and notes will
display on the next line instead.  This should cause output to
look nicer in the case there are longer test names.

Also instead of hardcoding lengths, they are now in a dict, so
if we want to do something smarter at some point it will be
easier.

Change-Id: Idd03c73d63d05f2161da45c3e8f01f1e1b67d45d