Clean up Logging
This commit cleans up logging in a couple of ways:

First - logging is now a global instance in the logging module.
To use it a module or test simply imports the logger from
lib.logging.  This saves the effort of having to call the
test_util function to get a logger object.

Secondly all logger calls are set to use format strings.  This
makes them look cleaner in a lot of cases where multiple string
concatenation was occurring.

Change-Id: Icfc42c674303a09ef05c4765e88013bba46c461f